CacheManager - Split static logics into a seperate logic instance class Allow the load of different sources of configuration with a loader class
