+++ Core/Config/Loaders/ICacheConfigLoader.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace PubComp.Caching.Core.Config.Loaders
+{
+    // TODO: Add XML comments
+    public interface ICacheConfigLoader
+    {
+        IList<ConfigNode> LoadConfig();
+    }
+}
